Lobby System MVP Complete
Players can now join a game client's lobby - which is automatically created when the game client launches.
A player client can only join a game client if the lobby id is correct and they have supplied a username.
If a player client is the first one to join, they are assigned as VIP.
If a VIP assigned player leaves the lobby, it is reassigned to another player if they exist.
If the game client leaves, all clients are informed of this leave event (CURRENTLY BUGGED - 'BroadcastDestroyLobby' not working - find out tomorrow zzz).